+++ Package.swift
@@ -4,8 +4,8 @@
 import PackageDescription
 import Foundation
 
-let commonVersion: Version = "24.12.0-beta.1"
-let coreMapsVersion: Version = "11.12.0-beta.1"
+let commonVersion: Version = "24.13.0-SNAPSHOT-04-23--05-20.git-ad48dc8"
+let coreMapsVersion: Version = "11.13.0-SNAPSHOT-04-23--05-20.git-ad48dc8"
 let turfVersion: Version = "4.0.0"
 
 let mapboxMapsPath: String? = nil