+++ Sources/MapboxMaps/Foundation/Extensions/Core/Vec2.swift
@@ -1,0 +1,8 @@
+import Foundation
+import MapboxCoreMaps
+
+extension Vec2 {
+    var point: CGPoint {
+        CGPoint(x: x, y: y)
+    }
+}